backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WebSockets
python-socketio==5.11.0
python-engineio==4.9.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# AI and File Processing
openai==1.10.0
PyPDF2==3.0.1
python-pptx==0.6.23
Pillow==10.2.0
pytesseract==0.3.10
python-docx==1.1.0

# Data Export
pandas==2.2.0
openpyxl==3.1.2
reportlab==4.0.9

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
qrcode[pil]==7.4.2
shortuuid==1.0.11

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

google-generativeai
bcrypt==3.2.2
